Add api to validate that animations were disabled by Test Butler
Fixes #1
