+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 Add api for granting runtime permissions on API 23+
 Reduce log spam from NoDialogActivityController
-Preserve original values of animation scales after using Test Butler
+Preserve original values of animation scales after using Test Butler [#25](https://github.com/linkedin/test-butler/issues/25)
+Expose TestButler.verifyAnimationsDisabled for apps that want to validate that animations were actually disabled and fail fast [#1](https://github.com/linkedin/test-butler/issues/1)
 
 ## Version 1.0.1
